:art: style: revamped project-card and deprecated project-card-square
