+++ src/feat/project/index.ts
@@ -2,7 +2,7 @@ export { default as ProjectLayout } from "./components/project-layout";
 export { default as ProjectArea } from "./components/project-area";
 export { default as ProjectGrid } from "./components/project-grid";
 export { default as ProjectDialog } from "./components/project-dialog";
-export { default as ProjectPlayingCard } from "./components/project-playing-card";
 export { default as ProjectCard } from "./components/project-card";
+export { default as ProjectCardSquare } from "./components/project-card-square";
 export { default as ProjectLayoutButton } from "./components/project-layout-button";
 export type { Project } from "./types";
